% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{CSI 300 Index Realized Measures}
\format{
An object with the following elements:
\describe{
\item{RV}{times series of daily realized volatility estimates}
\item{BPV}{times series of daily bi-power realized volatility estimates}
\item{JV}{time series of daily jump variation estimates}
\item{return}{time series of daily low-frequency returns}
}
}
\usage{
sample_data
}
\description{
This sample data set contains realized measures, such as realized volatility (RV),
bi-power realized volatility (BPV) and jump variation (JV) estimated from CSI 300 Index high-frequency data,
it also includes daily low-frequency log returns (return).
}
\keyword{datasets}
